Remove version & author from each subcommand
Currently every individual subcommand of youki has clap attributes giving
the youki version and authors.  That seems a bit redundant, since this
information is also on the top level Opts structure.

Having the tags on each does have an effect though: it means that, e.g.
"youki create --help" will give the version and author, as well as plain
"youki --help".  Likewise "youki create --version" will give a result as
well as "youki --version".

Including this is of limited value, though, since it can be easily obtained
from the top level command.  It also provides a barrier to splitting out
the frontend argument parsing from the rest of youki, so remove it.

Signed-off-by: David Gibson <david@gibson.dropbear.id.au>