Initial checkpoint support
This adds the first code to checkpoint a container. The checkpoint
command is name 'checkpointt' (with two 't's at the end) so that
container engines like Podman do not think to use this not yet finished
checkpoint restore implementation.

For Podman it is still necessary to tell CRIU that the network namespace
is external at least and restoring needs special handling to support
'--console-socket'.

Signed-off-by: Adrian Reber <areber@redhat.com>