Add tests to crates/libcgroups/src/v2/devices/controller.rs
In order to add these tests, I added `mockall` to the dev dependencies.

There are some things that could be cleaner, eg using mockall's `double`
or `cfg_if` but I did not want to bloat the release dependencies with extra crates.

This required some changes to the v2/devices/bpf module, introducing an
intermediate module `prog` to wrap the various `prog_*` functions inside.
Those functions were then renamed to remove the `prog_` prefix.
The addition of this intermediate module allowed us to use the `automock`
macro to automatically generate mocked functions. This is a useful
testing pattern that can be used in the future on the bpf module itself.

Signed-off-by: Christian Burke <cr0ax64@gmail.com>