Add more mocked unit tests for libcgroups in v2/devices/bpf.rs
This commit builds on previous work using the mockall library.
We add a mock module with signatures of functions we want to mock from
`libc` and `libbpf`, and then those mocks are used from the `bpf` module.

Tests for the `bpf_query` function required things like setting errors and
manipulating the content of c-arrays in the `returning` callback.

Signed-off-by: Christian Burke <cr0ax64@gmail.com>