Thaw a paused container in cgroup v1 when it is forcely deleted.
If force is not given, and cgroups is v1 and container is frozen, display
error saying could not be killed as it was forzen. If force option is given,
and cgroups is v1 and container is frozen, thaw it and send the kill signal.
If cgroups is v2 , nothing special needs to be done.

Fix: #1129

Signed-off-by: Chen Yiyang <cyyzero@qq.com>