Use /proc/self/uid_map to determine if rootless or not
When Podman starts a container in rootless, it first creates a user namespace
and then starts the container as root in it. Therefore, determining rootless
in youki with geteuid(2) is not sufficient.

Youki should look at /proc/self/uid_map to determine if it is rootless,
just like crun and runc.

Signed-off-by: Manabu Ori <manabu.ori@gmail.com>