refactored executor and executor manager
Refactored the executor and executor manager to be clone-able and more
ergonomics.

Removed the use of executor manager with a vector of executor. Instead
we favor composing executors into a new executor. Composing also allows
the implementer to be precise about the execution orders. The old
executor manager with its vector implementation is not explicit enough
about the order when multiple executor is involved.

Re-implement the executor to be a function pointer instead of the Fn
family traits. Traits will type erase and difficult to deal with when
implementing clone.

Refactored the wasm related exectors to use the new scheme.

Signed-off-by: yihuaf <yihuaf@unkies.org>